Move AppRouter
