+++ src/AppRouter.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import App from '../App';
+import App from './App';
 
 export default function AppRouter(props) {
   return (
