+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import AppRouter from './utils/AppRouter';
+import AppRouter from './AppRouter';
 import * as serviceWorker from './serviceWorker';
 import { HashRouter } from 'react-router-dom';
